HV-1991 Upgrade to JDK 17 as the baseline and drop JDK 11 compatibility
