+++ annotation-processor/src/test/java/org/hibernate/validator/ap/record/RecordWithInvalidConstraints.java
